requirements.txt
asgiref==3.4.1
beautifulsoup4==4.11.1
bootstrap4==0.1.0
click==8.0.4
Django==3.2.13
django-bootstrap4==21.1
django-filter==21.1
importlib-metadata==2.1.3
importlib-resources==5.4.0
joblib==1.1.0
nltk==3.6.7
numpy==1.19.5
pkg_resources==0.0.0
pytz==2022.1
regex==2022.3.15
scikit-learn==0.24.2
scipy==1.5.4
sklearn==0.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
threadpoolctl==3.1.0
tqdm==4.64.0
typing_extensions==4.1.1
zipp==3.6.0
